Crash in place inside ContentBlockingConfiguration
