+++ iOS/DuckDuckGo/AppConfiguration/ContentBlockingConfiguration.swift
@@ -22,20 +22,33 @@ import Core
 
 final class ContentBlockingConfiguration {
 
-    static func prepareContentBlocking() throws {
-        var thrownError: Error?
+    private let application: UIApplication
+
+    init(application: UIApplication = UIApplication.shared) {
+        self.application = application
+    }
+
+    func prepareContentBlocking() {
         ContentBlocking.shared.onCriticalError = {
-            do {
-                throw UIApplication.TerminationError.unrecoverableState
-            } catch {
-                thrownError = error
+            DispatchQueue.main.async {
+                self.alertAndTerminate()
             }
-        }
-        if let thrownError {
-            throw thrownError
         }
         // Explicitly prepare ContentBlockingUpdating instance before Tabs are created
         _ = ContentBlockingUpdating.shared
+    }
+
+    private func alertAndTerminate() {
+        let window: UIWindow
+        if let existingWindow = application.window {
+            window = existingWindow
+        } else {
+            window = UIWindow.makeBlank()
+            application.setWindow(window)
+        }
+
+        let alertController = CriticalAlerts.makePreemptiveCrashAlert()
+        window.rootViewController?.present(alertController, animated: true, completion: nil)
     }
 
 }
