+++ iOS/DuckDuckGo/UIWindowExtension.swift
@@ -1,0 +1,36 @@
+//
+//  UIWindowExtension.swift
+//  DuckDuckGo
+//
+//  Copyright © 2025 DuckDuckGo. All rights reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+import UIKit
+
+extension UIWindow {
+
+    static func makeBlank() -> UIWindow {
+        let window = UIWindow(frame: UIScreen.main.bounds)
+        window.backgroundColor = .white
+
+        let rootViewController = UIViewController()
+        rootViewController.view.backgroundColor = .white
+        window.rootViewController = rootViewController
+        window.makeKeyAndVisible()
+
+        return window
+    }
+
+}
